Removed all instances of unused classes and moved things around.
