--- svc-bip-api/src/main/java/gov/va/vro/bip/model/BipClaimResp.java
@@ -1,12 +1,0 @@
-package gov.va.vro.bip.model;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@JsonIgnoreProperties(ignoreUnknown = true) // BIP API can send messages
-public class BipClaimResp extends HasStatusCodeAndMessage {
-  private BipClaim claim;
-}
