--- svc-bip-api/src/main/java/gov/va/vro/bip/model/BipUpdateClaimStatusResp.java
@@ -1,16 +1,0 @@
-package gov.va.vro.bip.model;
-
-import lombok.Data;
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-@Data
-public class BipUpdateClaimStatusResp {
-  private boolean isSuccessful;
-  private String message;
-
-  public BipUpdateClaimStatusResp(boolean success, String msg) {
-    this.isSuccessful = success;
-    this.message = msg;
-  }
-}
