--- svc-bip-api/src/main/java/gov/va/vro/bip/model/CreateContentionReq.java
@@ -1,19 +1,0 @@
-package gov.va.vro.bip.model;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import lombok.Setter;
-
-import java.util.List;
-
-/**
- * Request to create contentions for a claim.
- *
- * @author warren @Date 11/14/22
- */
-@Getter
-@Setter
-@RequiredArgsConstructor
-public class CreateContentionReq {
-  private List<Contention> createContentions;
-}
