--- svc-bip-api/src/main/java/gov/va/vro/bip/model/UpdateContentionModel.java
@@ -1,18 +1,0 @@
-package gov.va.vro.bip.model;
-
-import lombok.Builder;
-import lombok.Getter;
-import lombok.Setter;
-
-/**
- * List of update contentions for a claim.
- *
- * @author greene @Date 9/24/23
- */
-@Getter
-@Setter
-@Builder
-public class UpdateContentionModel {
-  public long claimId;
-  private UpdateContentionReq updateContentions;
-}
