--- svc-bip-api/src/main/java/gov/va/vro/bip/model/UpdateContentionReq.java
@@ -1,19 +1,0 @@
-package gov.va.vro.bip.model;
-
-import lombok.Builder;
-import lombok.Getter;
-import lombok.Setter;
-
-import java.util.List;
-
-/**
- * Request to update contentions for a claim.
- *
- * @author warren @Date 11/14/22
- */
-@Getter
-@Setter
-@Builder
-public class UpdateContentionReq {
-  List<ExistingContention> updateContentions;
-}
