+++ svc-bip-api/src/main/java/gov/va/vro/bip/model/claim/BipClaim.java
@@ -1,4 +1,4 @@
-package gov.va.vro.bip.model;
+package gov.va.vro.bip.model.claim;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
