+++ svc-bip-api/src/main/java/gov/va/vro/bip/model/contentions/Contention.java
@@ -1,4 +1,4 @@
-package gov.va.vro.bip.model;
+package gov.va.vro.bip.model.contentions;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
@@ -57,5 +57,5 @@ public class Contention {
   private List<String> specialIssueCodes;
 
   @JsonProperty("associatedTrackedItems")
-  private List<TrackedItemAssociations> associatedTrackedItems;
+  private List<TrackedItemAssociation> associatedTrackedItems;
 }
