+++ svc-bip-api/src/main/java/gov/va/vro/bip/model/contentions/UpdateClaimContentionsRequest.java
@@ -4,7 +4,6 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import gov.va.vro.bip.model.BipPayloadRequest;
-import gov.va.vro.bip.model.ExistingContention;
 import lombok.Builder;
 import lombok.Getter;
