--- svc-bip-api/src/main/java/gov/va/vro/bip/service/ClaimProps.java
@@ -1,14 +1,0 @@
-package gov.va.vro.bip.service;
-
-import lombok.Getter;
-import lombok.Setter;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-
-/** @author warren @Date 2/7/23 */
-@Getter
-@Setter
-@ConfigurationProperties(prefix = "claim")
-public class ClaimProps {
-  private String specialIssue1;
-  private String specialIssue2;
-}
